fix test name
